speed up block memory transfers
With the clock set at 4MHz (measures at ~3.3MHz on the scope)
on an LPC MCU-LINK with 512byte packets, we see about 74KB/s
for the simple bulk reads.  Using auto increment in the CSW
and removing almost all MAP.TAR writes, we get up to 202KB/s.

It may be possible to get somewhat better efficiency by using
DAP_TransferBlock instead of DAP_Transfer, but I suspect that's
not going to have much impact on high speed USB devices like
this one.  On full speed devices, could be significant.